backend/requirements.txt
duckdb>=1.0.0
fastapi>=0.115.0
httpx>=0.27.0
numpy>=1.26.0
pandas>=2.1.0
pydantic>=2.7.0
pytest>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn[standard]>=0.30.0
